feat: add types file and simplify types
