+++ src/apis/products/productsController/index.ts
@@ -1,17 +1,17 @@
 import type { NextApiRequest, NextApiResponse } from "next"
-import { Data, ErrorData } from "@/pages/api/users"
 import { getProducts } from "../productsService"
+import { ResponseData, ResponseErrorData } from "@/types/types"
 
 // Handles HTTP GET request to retrieve products from database
 export const handleGetProducts = async (
   req: NextApiRequest,
-  res: NextApiResponse<Data | ErrorData>,
+  res: NextApiResponse<ResponseData | ResponseErrorData>,
 ) => {
   try {
-    const data: Data = await getProducts()
+    const data: ResponseData = await getProducts()
     res.status(200).json(data)
   } catch (error) {
-    const errorData: ErrorData = { message: 'Internal Server Error. Cannot GET products.' }
+    const errorData: ResponseErrorData = { message: 'Internal Server Error. Cannot GET products.' }
     res.status(500).json(errorData)
   }
 }
