+++ src/apis/users/usersService/index.ts
@@ -1,15 +1,9 @@
+import { User } from "@/types/types"
 import { getDataFromDb } from "../../../../db/utils/dbGet"
 import { addDataToDb } from "../../../../db/utils/dbPost"
 
-interface UserParams {
-  email: string;
-  password: string;
-  first_name?: string;
-  last_name?: string;
-}
-
 // Retrieves user from the database filtered by params
-export const getFilteredUser = async (params: UserParams) => getDataFromDb('dogstore_users', params)
+export const getFilteredUser = async (params: User) => getDataFromDb('dogstore_users', params)
 
 // Adds a user to the database
-export const addUser = async (params: UserParams) => addDataToDb('dogstore_users', params)
+export const addUser = async (params: User) => addDataToDb('dogstore_users', params)
