+++ src/components/ShopContent/utils/productFilter.ts
@@ -1,4 +1,4 @@
-import { Product } from "@/pages/shop"
+import { Product } from "@/types/types"
 
 export const productFilter = (productCategory: string, productsData: Product[]) => {
   if (productCategory === 'bestsellers') {
